+++ gusto_embedded/src/lib/config.ts
@@ -64,8 +64,8 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2025-06-15",
-  sdkVersion: "0.11.8",
+  sdkVersion: "0.11.9",
   genVersion: "2.801.0",
   userAgent:
-    "speakeasy-sdk/typescript 0.11.8 2.801.0 2025-06-15 @gusto/embedded-api",
+    "speakeasy-sdk/typescript 0.11.9 2.801.0 2025-06-15 @gusto/embedded-api",
 } as const;